FPLA-1774: added NoticeOfProceedingsTemplateDataGenerationService
