+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/NoticeOfProceedingsTemplateDataGenerationServiceTest.java
@@ -1,0 +1,238 @@
+package uk.gov.hmcts.reform.fpl.service;
+
+import com.google.common.collect.ImmutableList;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import uk.gov.hmcts.reform.fpl.model.CaseData;
+import uk.gov.hmcts.reform.fpl.model.Child;
+import uk.gov.hmcts.reform.fpl.model.ChildParty;
+import uk.gov.hmcts.reform.fpl.model.HearingBooking;
+import uk.gov.hmcts.reform.fpl.model.NoticeOfProceedings;
+import uk.gov.hmcts.reform.fpl.model.Orders;
+import uk.gov.hmcts.reform.fpl.model.common.Element;
+import uk.gov.hmcts.reform.fpl.model.common.JudgeAndLegalAdvisor;
+import uk.gov.hmcts.reform.fpl.model.docmosis.DocmosisNoticeOfProceeding;
+import uk.gov.hmcts.reform.fpl.service.config.LookupTestConfig;
+import uk.gov.hmcts.reform.fpl.service.time.Time;
+import uk.gov.hmcts.reform.fpl.utils.FixedTimeConfiguration;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.format.FormatStyle;
+import java.util.List;
+import java.util.UUID;
+
+import static java.util.Collections.emptyList;
+import static org.assertj.core.api.Assertions.assertThat;
+import static uk.gov.hmcts.reform.fpl.enums.JudgeOrMagistrateTitle.HIS_HONOUR_JUDGE;
+import static uk.gov.hmcts.reform.fpl.enums.JudgeOrMagistrateTitle.MAGISTRATES;
+import static uk.gov.hmcts.reform.fpl.enums.OrderType.CARE_ORDER;
+import static uk.gov.hmcts.reform.fpl.enums.OrderType.EDUCATION_SUPERVISION_ORDER;
+import static uk.gov.hmcts.reform.fpl.utils.CaseDataGeneratorHelper.createHearingBooking;
+import static uk.gov.hmcts.reform.fpl.utils.CaseDataGeneratorHelper.createPopulatedApplicants;
+import static uk.gov.hmcts.reform.fpl.utils.CaseDataGeneratorHelper.createPopulatedChildren;
+import static uk.gov.hmcts.reform.fpl.utils.DateFormatterHelper.formatLocalDateToString;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(classes = {NoticeOfProceedingsTemplateDataGenerationService.class})
+@ContextConfiguration(classes = {
+    JacksonAutoConfiguration.class, LookupTestConfig.class, HearingBookingService.class,
+    HearingVenueLookUpService.class, CaseDataExtractionService.class, FixedTimeConfiguration.class
+})
+class NoticeOfProceedingsTemplateDataGenerationServiceTest {
+
+    @Autowired
+    private Time time;
+
+    @Autowired
+    private NoticeOfProceedingsTemplateDataGenerationService noticeOfProceedingsTemplateDataGenerationService;
+
+    private LocalDate futureDate;
+    private List<Element<Child>> children;
+
+    @BeforeEach
+    void setup() {
+        futureDate = time.now().toLocalDate().plusDays(1);
+        children = createPopulatedChildren(time.now().toLocalDate());
+    }
+
+    @Test
+    void shouldApplySentenceFormattingWhenMultipleChildrenExistOnCase() {
+        CaseData caseData = initNoticeOfProceedingCaseData()
+            .children1(children)
+            .orders(Orders.builder()
+                .orderType(ImmutableList.of(CARE_ORDER)).build())
+            .noticeOfProceedings(NoticeOfProceedings.builder()
+                .judgeAndLegalAdvisor(createJudgeAndLegalAdvisor())
+                .proceedingTypes(emptyList())
+                .build())
+            .build();
+
+        DocmosisNoticeOfProceeding templateData = noticeOfProceedingsTemplateDataGenerationService
+            .getTemplateData(caseData);
+        assertThat(templateData.getChildrenNames()).isEqualTo("Bran Stark, Sansa Stark and Jon Snow");
+    }
+
+    @Test
+    void shouldNotApplySentenceFormattingWhenOnlyOneChildExistsOnCase() {
+        CaseData caseData = initNoticeOfProceedingCaseData()
+            .children1(ImmutableList.of(
+                Element.<Child>builder()
+                    .id(UUID.randomUUID())
+                    .value(Child.builder()
+                        .party(ChildParty.builder()
+                            .firstName("Bran")
+                            .lastName("Stark")
+                            .build())
+                        .build())
+                    .build()))
+            .orders(Orders.builder()
+                .orderType(ImmutableList.of(CARE_ORDER)).build())
+            .noticeOfProceedings(NoticeOfProceedings.builder()
+                .judgeAndLegalAdvisor(createJudgeAndLegalAdvisor())
+                .proceedingTypes(emptyList())
+                .build())
+            .build();
+
+        DocmosisNoticeOfProceeding templateData = noticeOfProceedingsTemplateDataGenerationService
+            .getTemplateData(caseData);
+        assertThat(templateData.getChildrenNames()).isEqualTo("Bran Stark");
+    }
+
+    @Test
+    void shouldFormatMagistrateFullNameWhenJudgeTitleIsSetToMagistrate() {
+        CaseData caseData = initNoticeOfProceedingCaseData()
+            .children1(children)
+            .noticeOfProceedings(NoticeOfProceedings.builder()
+                .judgeAndLegalAdvisor(JudgeAndLegalAdvisor.builder()
+                    .judgeTitle(MAGISTRATES)
+                    .judgeFullName("James Nelson")
+                    .build())
+                .proceedingTypes(emptyList())
+                .build())
+            .orders(Orders.builder()
+                .orderType(ImmutableList.of(CARE_ORDER)).build())
+            .build();
+
+        DocmosisNoticeOfProceeding templateData = noticeOfProceedingsTemplateDataGenerationService
+            .getTemplateData(caseData);
+        assertThat(templateData.getJudgeTitleAndName()).isEqualTo("James Nelson (JP)");
+    }
+
+    @Test
+    void shouldSetJudgeTitleAndNameToEmptyStringWhenJudgeTitleAndNameIsEmpty() {
+        CaseData caseData = initNoticeOfProceedingCaseData()
+            .children1(children)
+            .orders(Orders.builder()
+                .orderType(ImmutableList.of(CARE_ORDER)).build())
+            .noticeOfProceedings(NoticeOfProceedings.builder()
+                .judgeAndLegalAdvisor(JudgeAndLegalAdvisor.builder().build())
+                .proceedingTypes(emptyList())
+                .build())
+            .build();
+        DocmosisNoticeOfProceeding templateData = noticeOfProceedingsTemplateDataGenerationService
+            .getTemplateData(caseData);
+        assertThat(templateData.getJudgeTitleAndName()).isEqualTo("");
+    }
+
+    @Test
+    void shouldReturnFirstApplicantNameWhenMultipleApplicantsArePresent() {
+        CaseData caseData = initNoticeOfProceedingCaseData()
+            .children1(children)
+            .orders(Orders.builder()
+                .orderType(ImmutableList.of(CARE_ORDER)).build())
+            .noticeOfProceedings(NoticeOfProceedings.builder()
+                .judgeAndLegalAdvisor(createJudgeAndLegalAdvisor())
+                .proceedingTypes(emptyList())
+                .build())
+            .build();
+
+        DocmosisNoticeOfProceeding templateData = noticeOfProceedingsTemplateDataGenerationService
+            .getTemplateData(caseData);
+        assertThat(templateData.getApplicantName()).isEqualTo("Bran Stark");
+    }
+
+    @Test
+    void shouldMapCaseDataPropertiesToTemplatePlaceholderDataWhenCaseDataIsFullyPopulated() {
+        CaseData caseData = initNoticeOfProceedingCaseData()
+            .children1(children)
+            .noticeOfProceedings(NoticeOfProceedings.builder()
+                .judgeAndLegalAdvisor(createJudgeAndLegalAdvisor())
+                .proceedingTypes(emptyList())
+                .build())
+            .orders(Orders.builder()
+                .orderType(ImmutableList.of(
+                    CARE_ORDER,
+                    EDUCATION_SUPERVISION_ORDER
+                )).build())
+            .build();
+
+        DocmosisNoticeOfProceeding templateData = noticeOfProceedingsTemplateDataGenerationService
+            .getTemplateData(caseData);
+
+        DocmosisNoticeOfProceeding expectedData = DocmosisNoticeOfProceeding.builder()
+            .courtName("Family Court")
+            .familyManCaseNumber("123")
+            .applicantName("Bran Stark")
+            .orderTypes("Care order, Education supervision order")
+            .childrenNames("Bran Stark, Sansa Stark and Jon Snow")
+            .hearingDate(formatLocalDateToString(futureDate, FormatStyle.LONG))
+            .hearingVenue("Crown Building, Aberdare Hearing Centre, Aberdare, CF44 7DW")
+            .preHearingAttendance("8:30am")
+            .hearingTime("9:30am - 11:30am")
+            .judgeTitleAndName("His Honour Judge Samuel Davidson")
+            .legalAdvisorName("John Bishop")
+            .todaysDate(formatLocalDateToString(time.now().toLocalDate(), FormatStyle.LONG))
+            .crest("[userImage:crest.png]")
+            .courtseal("[userImage:familycourtseal.png]")
+            .build();
+
+        assertThat(templateData).isEqualToComparingFieldByField(expectedData);
+    }
+
+    private JudgeAndLegalAdvisor createJudgeAndLegalAdvisor() {
+        return JudgeAndLegalAdvisor.builder()
+            .judgeTitle(HIS_HONOUR_JUDGE)
+            .judgeLastName("Samuel Davidson")
+            .legalAdvisorName("John Bishop")
+            .build();
+    }
+
+    private List<Element<HearingBooking>> createHearingBookings() {
+        return ImmutableList.of(
+            Element.<HearingBooking>builder()
+                .id(UUID.randomUUID())
+                .value(createHearingBooking(
+                    LocalDateTime.of(futureDate, LocalTime.of(9, 30)),
+                    LocalDateTime.of(futureDate, LocalTime.of(11, 30))))
+                .build(),
+            Element.<HearingBooking>builder()
+                .id(UUID.randomUUID())
+                .value(createHearingBooking(
+                    LocalDateTime.of(futureDate, LocalTime.of(12, 30)),
+                    LocalDateTime.of(futureDate, LocalTime.of(13, 30))))
+                .build(),
+            Element.<HearingBooking>builder()
+                .id(UUID.randomUUID())
+                .value(createHearingBooking(
+                    LocalDateTime.of(futureDate, LocalTime.of(15, 30)),
+                    LocalDateTime.of(futureDate, LocalTime.of(16, 0))))
+                .build()
+        );
+    }
+
+    private CaseData.CaseDataBuilder initNoticeOfProceedingCaseData() {
+        return CaseData.builder()
+            .caseLocalAuthority("example")
+            .familyManCaseNumber("123")
+            .applicants(createPopulatedApplicants())
+            .hearingDetails(createHearingBookings());
+    }
+}
